+++ test/simple_graphql_mock_test.dart
@@ -1,114 +1,66 @@
-import 'package:graphql/client.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:simple_graphql/simple_graphql.dart';
-import 'package:simple_graphql/src/simple_graphql_mock.dart';
 import 'package:test/test.dart';
 
-// void main() {
-
-// }
-
-// Copyright (c) 2012, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-import 'dart:convert';
-
-import 'package:http/http.dart' as http;
-import 'package:http/testing.dart';
-import 'package:test/test.dart';
+class SimpleGraphQlMock extends Mock implements SimpleGraphQl {}
 
 void main() {
-  const apiUrl = 'https://api.myapi.example/graphql';
+  setUpAll(() {
+    registerFallbackValue(SimpleGraphQlMock());
+  });
 
   group('SimpleGraphQlMock handles queries', () {
-    test('should handles a query with a success response data', () async {
-      final responseExpected = {'field1': 'value1', 'field2': 'value2'};
-      final client = SimpleGraphQlMock(
-        apiUrl: apiUrl,
-        handler: (operation, variables, token) {
-          return MockQueryResult.test(data: responseExpected);
-        },
+    test('should login successfully using correct credentials', () async {
+      const query = r'''
+        query Login($username: String, $password, String) { 
+          login(username: $username, password: $password) {
+            success
+            token
+          }
+        }''';
+
+      final client = SimpleGraphQlMock();
+
+      final variables = {'username': 'john_doe', 'password': '12345'};
+
+      when(
+        () => client.query<Map<String, dynamic>>(
+          query: any(named: 'query'),
+          variables: any(named: 'variables'),
+        ),
+      ).thenAnswer(
+        (_) async =>
+            <String, dynamic>{'success': true, 'token': 'test_eyJhb....sw5c'},
       );
-      final response = await client.query(
-        query: 'query ExampleQuery { data }',
-        resultBuilder: (data) => data,
+
+      final response = await client.query<Map<String, dynamic>>(
+        variables: variables,
+        query: query,
       );
+
+      final captured = verify(
+        () => client.query<Map<String, dynamic>>(
+          query: captureAny(named: 'query'),
+          variables: captureAny(named: 'variables'),
+        ),
+      ).captured.toList();
+
+      // captured List contains the variables given.
+      expect(captured, contains(variables));
+
+      final capturedVariables = captured[1] as Map<String, dynamic>;
+
+      // captured Map contains the 'username' and 'password' variables
+      expect(capturedVariables.containsKey('username'), true);
+      expect(capturedVariables.containsKey('password'), true);
 
       expect(
         response,
-        responseExpected,
-      );
-    });
-
-    test('should handles a query with an Exception', () async {
-      final client = SimpleGraphQlMock(
-        apiUrl: apiUrl,
-        handler: (operation, variables, token) {
-          return MockQueryResult.test(
-            exception: OperationException(),
-          );
+        {
+          'success': true,
+          'token': 'test_eyJhb....sw5c',
         },
       );
-
-      expect(
-        () => client.query(
-          query: 'query ExampleQuery { data }',
-          resultBuilder: (data) => data,
-        ),
-        throwsA(isA<SimpleGqlException>()),
-      );
-    });
-
-    test('should handles a query with a GraphQlError with message', () async {
-      const errorMessageExpected = 'Ups, sorry';
-      final client = SimpleGraphQlMock(
-        apiUrl: apiUrl,
-        handler: (operation, variables, token) {
-          return MockQueryResult.test(
-            exception: OperationException(
-              graphqlErrors: [
-                const GraphQLError(message: errorMessageExpected),
-              ],
-            ),
-          );
-        },
-      );
-
-      try {
-        await client.query(
-          query: 'query ExampleQuery { data }',
-          resultBuilder: (data) => data,
-        );
-      } catch (e) {
-        expect(e, isA<SimpleGqlException>());
-        if (e is SimpleGqlException) {
-          expect(e.message, errorMessageExpected);
-        }
-      }
     });
   });
-
-  // test('handles a streamed request', () async {
-  // final client = MockClient.streaming((request, bodyStream) async {
-  //     final bodyString = await bodyStream.bytesToString();
-  //     final stream =
-  //         Stream.fromIterable(['Request body was "$bodyString"'.codeUnits]);
-  //     return http.StreamedResponse(stream, 200);
-  //   });
-
-  //   final uri = Uri.http('example.com', '/foo');
-  //   final request = http.Request('POST', uri)..body = 'hello, world';
-  //   final streamedResponse = await client.send(request);
-  //   final response = await http.Response.fromStream(streamedResponse);
-  //   expect(response.body, equals('Request body was "hello, world"'));
-  // });
-
-  // test('handles a request with no body', () async {
-  //   final client = MockClient((_) async => http.Response('you did it', 200));
-
-  //   expect(
-  //     await client.read(Uri.http('example.com', '/foo')),
-  //     equals('you did it'),
-  //   );
-  // });
 }