+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.0
+* Added testing support to the library. 
+  - You can now write unit tests using the `SimpleGraphQlMock` class, similar to `MockClient` from [`package:http`](https://pub.dev/packages/http)
 ## 0.1.0+2
 * Documentation update.
   - The package usage have been documented.